# docs-site/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Parcel Verification Service",
    "version": "1.0.0",
    "description": "BhoomiPe Parcel Verification Service — accepts a land parcel for verification, moves it through a state machine involving an external registry partner, and keeps a full audit trail of every state change.\n\n**Documents:** single attach takes a file picker plus a `document_type` dropdown, and the file is optional (omit it to record metadata only). Batch attach has one file picker per document type — upload each file under the field named after its type, in any combination; the field name is what classifies the document."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current version"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "INVALID_TRANSITION"
              },
              "message": {
                "type": "string",
                "example": "Cannot move from 'verified' to 'under_verification'."
              },
              "request_id": {
                "type": "string",
                "example": "9c1f2e3a-..."
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "document_type": {
            "type": "string"
          },
          "original_file_name": {
            "type": "string"
          },
          "content_type": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "document_type",
          "original_file_name",
          "content_type",
          "size_bytes",
          "uploaded_at"
        ]
      },
...
      "BatchDocumentsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Document"
            }
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "data",
          "count"
        ]
      }
    },
...
  },
  "paths": {
...
    "/parcels/{id}/documents": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Attach a document to a parcel. File is optional — omit or send empty to record metadata only (local disk when bytes are present).",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3f1c2a4-1234-4a9b-8b1e-000000000001"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional PDF/PNG/JPEG (max 10MB). Omit or empty for metadata-only."
                  },
                  "document_type": {
                    "type": "string",
                    "enum": [
                      "sale_deed",
                      "record_of_rights",
                      "encumbrance_certificate",
                      "owner_id_proof",
                      "other"
                    ],
                    "description": "Document classification.",
                    "example": "sale_deed"
                  }
                },
                "required": [
                  "document_type"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "400": {
            "description": "Bad document_type, or file content doesn't match declared mime type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No parcel with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/parcels/{id}/documents/batch": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Attach multiple documents in one request (all-or-nothing). Upload each file under the field named after its document type; max 10 files total. For a metadata-only record, use the single attach endpoint.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3f1c2a4-1234-4a9b-8b1e-000000000001"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "sale_deed": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "maxItems": 10,
                    "description": "Files to file as 'sale_deed'. PDF/PNG/JPEG, max 10MB each."
                  },
                  "record_of_rights": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "maxItems": 10,
                    "description": "Files to file as 'record_of_rights'. PDF/PNG/JPEG, max 10MB each."
                  },
                  "encumbrance_certificate": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "maxItems": 10,
                    "description": "Files to file as 'encumbrance_certificate'. PDF/PNG/JPEG, max 10MB each."
                  },
                  "owner_id_proof": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "maxItems": 10,
                    "description": "Files to file as 'owner_id_proof'. PDF/PNG/JPEG, max 10MB each."
                  },
                  "other": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "maxItems": 10,
                    "description": "Files to file as 'other'. PDF/PNG/JPEG, max 10MB each."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "All documents attached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDocumentsResponse"
                }
              }
            }
          },
          "400": {
            "description": "No files attached, more than 10 files, unknown field name, or a file whose content doesn't match its declared mime type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No parcel with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}